backend/api_gateway/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
httpx
python-jose
sqlalchemy
psycopg2-binary
pydantic[email]
email-validator